Play button on deepzoom webms
