+++ app/src/main/res/layout/fragment_deepzoom.xml
@@ -16,8 +16,20 @@
         android:indeterminate="true"
         android:visibility="invisible" />
 
-    <uk.co.senab.photoview.PhotoView xmlns:android="http://schemas.android.com/apk/res/android"
-        android:id="@+id/deepzoom_photoview"
+    <FrameLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent" />
+        android:layout_height="match_parent">
+        <uk.co.senab.photoview.PhotoView xmlns:android="http://schemas.android.com/apk/res/android"
+            android:id="@+id/deepzoom_photoview"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent" />
+
+        <ImageView
+            android:id="@+id/deepzoom_media_play_button"
+            android:layout_width="48dp"
+            android:layout_height="48dp"
+            android:layout_gravity="center_horizontal|center_vertical"
+            android:src="@drawable/ic_play_circle_outline_black_48dp"
+            android:visibility="gone"/>
+    </FrameLayout>
 </LinearLayout>
